fastapi-image-server/pyproject.toml
[project]
name = "fastapi-image-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "transformers>=4.55.4",
    "uvicorn>=0.35.0",
]
